Add JSON Web Token authentication
Add authentication using JSON Web Tokens (JWT). Users can make
requests using signed tokens that can in the future be created by
various authentication providers, including custom ones built and
hosted by the user.

The claims that are currently used are:
- `sub`: Subject. A unique id of the user.
- `exp`: Expiration Time.
- `admin`: Administrator flag (private claim name). This will give a
  permission to do anything, including mutating schemas, etc.

The authentication scheme allows the user to have one or more secrets
for their app and will try to verify the signature with each key until
it succeeds.

On successful verification it adds a `auth.credentials`
property including `userID` and `isAdmin` fields to the request
object. On failed authentication it responds with 401.

More on JWTs:
- https://auth0.com/blog/2014/12/02/using-json-web-tokens-as-api-keys/
- http://self-issued.info/docs/draft-ietf-oauth-json-web-token.html (RFC)
- http://jwt.io/ (debugger)