Implement social login
Add authentication using Facebook, GitHub, Google and Twitter.

For authentication, Bell is used, but we use its internals directly
instead of using it as a plugin in a normal way. This is because Bell
expects configuration to be static, but we need to fetch it from the
database, according to which app we're in.

The configuration for each authentication provider is stored using the
type ReindexAuthenticationProvider.

The authentication flow happens as follows:

1. Client library opens a popup to <REINDEX_URI>/auth/<provider>
2. The page sets state in a cookie and redirects to the OAuth provider
3. User enters login details and is redirected back to
  <REINDEX_URI>/auth/<provider>
4. We try to find a user with the matching 3rd party ID, and if not
  found create a new user.
5. The resulting HTML page passes a new JWT, the user or possible
errors from the popup to the client library.