Fix the mass breakage of tests
Awaiting on an Immutable.Map with `await*` does not await on all the
values of the map, because the iterator of a Map is an iterator of
(key, value) entries. Thus the code creating tables did not await
for all the promises, but continued immediately.

I don't know how this worked in the older RethinkDB (maybe the client
ran the commands sequentially or something), but with this fixed,
this runs on 2.1.1 as well.