Fix social login regressions
1) `clientID` had been renamed to `clientId` in the GraphQL type of the
authentication provider without changing it in the code that uses
authentication providers. Fix the code.

2) Token created by social login included the user ID in the raw
`{ type: 'User', value: <uuid> }` format. The permission checking code
expects it in the `<uuid>` format. Encode the user ID to opaque ID when
creating the token and decode the ID to `<uuid>` when reading the
token.