Add schema validator and migration builders
The migration pipeline consists of 3 parts:
* Validator
* Migration builder
* Migration executor

Schema validator validates the types of a schema before migration is created
from it. Migration builders take two versions of a schema and generate a list
of commands to migrate from one to the other. These commands form a schema
migration that will be executed by a migration executor that will make the
actual changes to the database.

Note:
Altering the type of a field or adding/removing an interface of a type
would require the data to be converted first. We sidestep this problem
by only having destructive operations for those types of changes.

- If interfaces have been added/removed, drop and recreate the type.
- If type of a field has been altered, drop and recreate the field.

If data should be preserved, the user should add a new field with
a different name.