Track all admin queries
This is useful because it lets us see in Intercom which users are
actively using the product.

At first I only tracked calls to `migrate` because I thought tracking
everything would add a lot of noise. However, we can just track queries
made as the admin user, as that typically means queries from CLI or
GraphiQL – not the queries made in apps of our users.

Also moved the tracking to setImmediate – tracking should not add any
slowdown to queries.