Fix non-null list fields
A schema with a non-null List field broke the input object
generator because GraphQLList wrapped in GraphQLNonNull was not
handled correctly. Pushing a schema like that resulted in the app
constantly returning Internal Server Error, because the schema could
not be created.

This commit fixes this by making the input object generator properly
unwrap the field type all the way through and then also wrap it back
again correctly.