Fix the broken npm run list-apps command
list-apps does not define a hostname, so it was broken when
RethinkDBClient expected a hostname. Made the hostname optional to fix
it.

In order to avoid accidentally making queries to the default database
(RethinkDB client uses `test` by default), added an assertion that
checks the hostname is still defined when using queries that require
it.