Add support for multiple database clusters
This commit implements support for arbitrary number of database
clusters. The database clusters are defined in 'database.clusters'
setting, which is a JSON map from cluster name to cluster config.

All apps, along with their domain names and database/cluster names
are stored in a single database (specified with 'database.adminCluster'
and 'database.adminDatabase' settings). This database is created like
any Reindex app and the schema defined in ReindexSchema.json.

Also updated the create-user script:

- Prompt for everything. I've found that this is much more convenient
  in practice than dealing with the positional arguments.
- Provide sensible defaults.
- Add support for multiple clusters.