Fix email formatting
Intercom doesn't support markdown or html in the body of messages sent
using the API. Remove formatting from the message to make it look
alright.

I have contacted Intercom support about this limitation and at the
the moment there isn't a good way to send messages like this with
Intercom.

Looks like we need to eventually use an actual transactional email
service for sending these emails, but this should work for now.