Fix bug with many-to-many connection-based permissions
When we are checking connection through permissions, we are making a 'stub object'
as the object we check, which is the object that only has the relevant connection
for which it is checked. This caused problems with hasPermission because our
hasPermission for many-to-many and many-to-one connections relied on checking through
object id, not through the connection itself. This PR adds a special case, for when
the id is not available.