Fix error from attempting to delete empty type by always creating them
So MongoDB collections are implicit, you create them on first access.
But then deleting a non-existant things is an error. However in 99%
of our use cases we do create collection, cause (eg) indexes. But in
that 1% user does delete just an empty type and get an error.

This makes collection creation explicit.