Handle closed connections
MongoClient auto reconnects connections by default, with 30 retries and 1000ms

delay between retries. If this fails, the Db is destroyed, which is the most

likely explanation for why we've seen apps get stuck trying to repeatedly use a

closed connection.



When a Db is destroyed we now take it out of the connection cache, so next

requests will get a fresh connection (if the connectivity has been restored),

instead of trying to reuse the destroyed connection pool.